roachtest,admission: assume ownership of adjacent tests
Release note: None
