+++ pkg/cmd/roachtest/registry/owners.go
@@ -20,6 +20,7 @@ const (
 	OwnerDisasterRecovery Owner = `disaster-recovery`
 	OwnerCDC              Owner = `cdc`
 	OwnerKV               Owner = `kv`
+	OwnerAdmissionControl Owner = `admission-control`
 	OwnerMultiRegion      Owner = `multiregion`
 	OwnerObsInf           Owner = `obs-inf-prs`
 	OwnerServer           Owner = `server` // not currently staffed
