+++ pkg/cmd/roachtest/tests/BUILD.bazel
@@ -8,6 +8,9 @@ go_library(
         "acceptance.go",
         "activerecord.go",
         "activerecord_blocklist.go",
+        "admission_control.go",
+        "admission_control_multi_store_overload.go",
+        "admission_control_tpcc_overload.go",
         "allocator.go",
         "alterpk.go",
         "asyncpg.go",
@@ -87,7 +90,6 @@ go_library(
         "network.go",
         "nodejs_postgres.go",
         "orm_helpers.go",
-        "overload_tpcc_olap.go",
         "parsing_helpers.go",
         "pebble_write_throughput.go",
         "pebble_ycsb.go",
