+++ pkg/cmd/roachtest/tests/registry.go
@@ -17,6 +17,7 @@ func RegisterTests(r registry.Registry) {
 	registerAWSDMS(r)
 	registerAcceptance(r)
 	registerActiveRecord(r)
+	registerAdmission(r)
 	registerAllocator(r)
 	registerAlterPK(r)
 	registerAsyncpg(r)
@@ -65,7 +66,6 @@ func RegisterTests(r registry.Registry) {
 	registerKVBench(r)
 	registerKVContention(r)
 	registerKVGracefulDraining(r)
-	registerKVMultiStoreWithOverload(r)
 	registerKVQuiescenceDead(r)
 	registerKVRangeLookups(r)
 	registerKVScalability(r)
@@ -82,7 +82,6 @@ func RegisterTests(r registry.Registry) {
 	registerMultiTenantUpgrade(r)
 	registerNetwork(r)
 	registerNodeJSPostgres(r)
-	registerOverload(r)
 	registerPebbleWriteThroughput(r)
 	registerPebbleYCSB(r)
 	registerPgjdbc(r)
